requirements.txt
psycopg2
simple-chalk
python-dotenv